Borrar y añadir objetivo, principio y propuesta Asignar presupuesto propuesta

+++ borrarobjetivo.php
@@ -1,6 +1,4 @@
      <?php
- if(isset($_POST['item2']))
-{
 
     // checks and alerts
     // ...
@@ -9,13 +7,19 @@
 
         $idObjetivo = $_POST['item2'];
         $mysqli = mysqli_connect("localhost","root","root", "test");
-        $query = "DELETE FROM `objetivos` WHERE `objetivos`.`id`=".$idObjetivo;
+        $proyectos = "SELECT * FROM `proyectos` WHERE `proyectos`.`alineadoObj` = ".$idObjetivo;
+        $total = mysql_num_rows($proyectos);
+        error_log("Failed to connect to database! ".$total." ". $idObjetivo , 0);
+        if ($total==0){
+        	 $query = "DELETE FROM `objetivos` WHERE `objetivos`.`id`= ".$idObjetivo;
+        	}
+       
         
 if (mysqli_query($mysqli, $query)== FALSE)   {
      echo "Error: " . $query . "<br>" . mysqli_error($mysqli);
 }
 mysqli_close($mysqli);
 }
-header('Location: objetivos.php');
-exit;
+//header('Location: objetivos.php');
+//exit;
 ?>
